pyproject.toml
[project]
name = "registry-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core API dependencies
    "aiohttp==3.12.15",
    "aws-xray-sdk==2.14.0",
    "bcrypt==4.1.3",
    "boto3==1.34.144",
    "botocore==1.34.144",
    "fastapi==0.111.0",
    "mangum==0.17.0",
    "pydantic==2.10.3",
    "pyjwt==2.8.0",
    "python-json-logger==2.0.7",
    "python-multipart==0.0.20",
    "requests==2.32.4",
    # Essential dependencies for Lambda deployment
    "email-validator==2.2.0",
    "python-dateutil==2.9.0.post0",
    "starlette==0.37.2",
    "typing-extensions==4.14.1",
    "urllib3>=1.25.4,<2.0.0",
    "wrapt==1.17.2",
    # Testing dependencies (production needed)
    "moto==5.0.9",
    "pytest==8.2.2",
    "pytest-mock==3.14.0",
    "pytest-asyncio>=1.1.0",
]

[dependency-groups]
dev = [
    # Development and testing tools
    "black==25.1.0",
    "flake8==7.3.0",
    "pytest-asyncio==1.1.0",
    
    # Additional development dependencies
    "httpx==0.28.1",
    "pandas==2.3.1",
    "openpyxl==3.1.5",
    "pyyaml==6.0.2",
    
    # FastAPI development tools
    "fastapi-cli==0.0.8",
    "uvicorn==0.35.0",
    
    # Rich formatting and CLI tools
    "rich==14.0.0",
    "rich-toolkit==0.14.8",
    "typer==0.16.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
